+++ packages/sdks/src/blocks/accordion/accordion.lite.tsx
@@ -177,16 +177,18 @@ export default function Accordion(props: AccordionProps) {
               class={state.getAccordionTitleClassName(index)}
               style={{
                 ...state.accordionTitleStyles,
-                width: props.grid ? props.gridRowWidth : undefined,
-                ...(useTarget({
-                  reactNative: {},
-                  default: {
-                    order:
-                      state.openGridItemOrder !== null
-                        ? convertOrderNumberToString(index)
-                        : convertOrderNumberToString(index + 1),
-                  },
-                }) as any),
+                ...(props.grid && {
+                  width: props.gridRowWidth,
+                  ...(useTarget({
+                    reactNative: {},
+                    default: {
+                      order:
+                        state.openGridItemOrder !== null
+                          ? convertOrderNumberToString(index)
+                          : convertOrderNumberToString(index + 1),
+                    },
+                  }) as any),
+                }),
               }}
               data-index={index}
               onClick={() => state.onClick(index)}