+++ packages/sdks/src/components/blocks/blocks-wrapper.lite.tsx
@@ -46,6 +46,15 @@ export default function BlocksWrapper(props: BlocksWrapperProps) {
         .filter(Boolean)
         .join(' ');
     },
+    get dataPath() {
+      if (!props.path) {
+        return undefined;
+      }
+      const pathPrefix = 'component.options.';
+      return props.path.startsWith(pathPrefix)
+        ? props.path
+        : `${pathPrefix}${props.path || ''}`;
+    },
     onClick() {
       if (isEditing() && !props.blocks?.length) {
         window.parent?.postMessage(
@@ -53,7 +62,7 @@ export default function BlocksWrapper(props: BlocksWrapperProps) {
             type: 'builder.clickEmptyBlocks',
             data: {
               parentElementId: props.parent,
-              dataPath: props.path,
+              dataPath: state.dataPath,
             },
           },
           '*'
@@ -67,7 +76,7 @@ export default function BlocksWrapper(props: BlocksWrapperProps) {
             type: 'builder.hoverEmptyBlocks',
             data: {
               parentElementId: props.parent,
-              dataPath: props.path,
+              dataPath: state.dataPath,
             },
           },
           '*'
@@ -84,10 +93,12 @@ export default function BlocksWrapper(props: BlocksWrapperProps) {
            * React Native strips off custom HTML attributes, so we have to manually set them here
            * to ensure that blocks are correctly dropped into the correct parent.
            */
-          props.path &&
-            blocksWrapperRef.setAttribute('builder-path', props.path);
-          props.parent &&
+          if (state.dataPath) {
+            blocksWrapperRef.setAttribute('builder-path', state.dataPath);
+          }
+          if (props.parent) {
             blocksWrapperRef.setAttribute('builder-parent-id', props.parent);
+          }
         }
       },
       default: () => {},
@@ -98,7 +109,7 @@ export default function BlocksWrapper(props: BlocksWrapperProps) {
     <props.BlocksWrapper
       ref={blocksWrapperRef}
       class={state.className}
-      builder-path={props.path}
+      builder-path={state.dataPath}
       builder-parent-id={props.parent}
       {...useTarget({
         reactNative: { dataSet: { class: state.className } },