+++ packages/sdks/output/qwik/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Builder.io Qwik SDK Changelog (@builder.io/sdk-qwik)
 
+## 0.16.22
+
+### Patch Changes
+
+- efa4798: Fix: accordion block order of items and visual editing empty blocks
+
+## 0.16.21
+
+### Patch Changes
+
+- a25a805: Fix: build to work in preview mode
+
 ## 0.16.20
 
 ### Patch Changes