+++ packages/sdks/output/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Builder.io React SDK Changelog (@builder.io/sdk-react)
 
+## 2.0.29
+
+### Patch Changes
+
+- efa4798: Fix: accordion block order of items and visual editing empty blocks
+
 ## 2.0.28
 
 ### Patch Changes